fix some minor Coverity issues
CID 92051: leak on errors

CID 90759: leak on errors

CID 90782: useless check

CID 90756: possible leak

CID 90739 90740: integer overflow, possibly memory corruption on error

+++ src/tds/net.c
@@ -887,7 +887,7 @@ tds7_get_instance_ports(FILE *output, struct addrinfo *addr)
 	int retval;
 	TDS_SYS_SOCKET s;
 	char msg[16*1024];
-	size_t msg_len = 0;
+	int msg_len = 0;
 	int port = 0;
 	char ipaddr[128];
 
@@ -1022,7 +1022,7 @@ tds7_get_instance_port(struct addrinfo *addr, const char *instance)
 	int retval;
 	TDS_SYS_SOCKET s;
 	char msg[1024];
-	size_t msg_len;
+	int msg_len;
 	int port = 0;
 	char ipaddr[128];
